refactor(confirm): delegate calculations to FastExecutionRequest
PendingArb.max_contracts() and profit_cents() now delegate to
FastExecutionRequest methods, eliminating duplicate calculation logic.

Also updated ConfirmationRecord.max_contracts from i64 to u16 for
type consistency.

+++ controller/src/confirm_log.rs
@@ -40,7 +40,7 @@ pub struct ConfirmationRecord {
     pub yes_price_cents: u16,
     pub no_price_cents: u16,
     pub profit_cents: i16,
-    pub max_contracts: i64,
+    pub max_contracts: u16,
     pub detection_count: u32,
     pub kalshi_url: String,
     pub poly_url: String,
